requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-core>=0.2.0
langgraph>=0.2.0
langchain-community>=0.2.0
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
notion-client>=2.2.1
todoist-api-python>=2.1.3
requests>=2.31.0
hijri-converter>=2.3.1
Pillow>=10.3.0
python-dotenv>=1.0.1
pydantic>=2.7.0
aiohttp>=3.9.5
rich>=13.7.1
SpeechRecognition>=3.10.4
